add class archivo

+++ src/data/models/Atraccion.java
@@ -11,7 +11,7 @@ public class Atraccion extends Oferta {
 	private static final List<String> LISTA_TIPOS = new ArrayList<String>(
 			Arrays.asList("Aventura", "Degustacion", "Paisaje"));
 
-	public Atraccion(String nombre, int precioFinal, float duracion, int cupos, String tipo) throws AtraccionExcepcion {
+	public Atraccion(String nombre, int precioFinal, double duracion, int cupos, String tipo) throws AtraccionExcepcion {
 
 		super(nombre, duracion, tipo);
